Docs in cinderlcient were never actually updated.
Not only were the docs and release info here not being
maintainted properly, but the examples that were given were
not correct.

This change is an attempt to set us on the right path going
forward and includes info on how to import the API and history
info for the upcoming release/push to PyPi.

Fixes bug: 1156994

Change-Id: I169b1c01380ef653a6a25eb8946b9a06d6419e62