Catch KeyboardInterrupt
When a user Ctrl-C's the client, this will cause a short message to
be printed rather than an ugly backtrace.

This will not catch all instances of this (can still Ctrl-C before
reaching this code, etc.), but it should catch most of them, and those
for long-running sessions.

The specific message was chosen to match one recently added to the
glance client.

Change-Id: Ib748835ad1f94fb901b4941dc5e3797b719e197f