Fixing erroneous clearing of test callstack
This modification addresses Bug #1210292, specifically addressing
the case where the test framework callstack is cleared when no
asserts are raised in assert_called_anytime. This test method
verifies that a specific API method call was made at any time
during a test. If a test makes multiple calls to this method, and
the first call yields no assertions, the subsequent calls will fail
because the first assert in the method verifies that the callstack
is non-empty. The error itself likely exists because
assert_called_anytime is not currently used in the
python-cinderclient testing framework.

The fix involves deleting the last line in the method that clears
the callstack.

Fixes: bug 1210292

Change-Id: I8b7b740957841a328b2f0ca888190f758cbdd234