Add ability to create volume from image by image name
Follow on change from cinder code to all specifying an
image by either its name or ID for creating a volume.

This patch adds the --image option that will take either
ID or name as input. The service then will attempt to
locate the image to use and make sure there are not
multiple images with the same name.

If both --image-id and --image are used it will favor
the more explicit ID.

DocImpact
Closes-Bug: 1377823

Change-Id: I77d58258144970fac47da6da6b7870f2931e1823