reset-state should warn that it is DB only
The python-cinderclient command 'reset-state' will only change the state in
the Cinder database, without affecting any connection/export of a volume from
the driver backend to a Nova compute host/instance.
The Doc string should explicitly state that this is the case, and Warn that it
might result in an unusable volume, i.e. a volume is set to 'available' in the
Cinder DB, but is still attached and therefore cannot be attached to another
instance, nor detached.

Change-Id: Iee66570dbb93776490ae8fb663a7f835dc496bc2
Closes-Bug: 1417273