Fix volume_transfers import in v2
Fixing the import of volume_transfers in v2 client.
It was importing volume_transfers from v1.
Even though at this moment the imported files are identical, in the
future volume_transfers in v2 may change, and v2 client should use the
v2 volume_transfers.
Since the intent is for v1 to be deprecated in the future, we must be
careful with these little details.

Closes-Bug: #1420238

Change-Id: I41c385de32ccf66d23002a8e369848016fa2c790