Fix outdated URLs and some minor fixes
A minor commit.
Credit to Jeremy Stanley for findind out this in other projects.

The developer documentation changed. It is now at infra manual, so now
the CONTRIBUTING.md was updated accordingly.

Fixed some other things, like a line break which was not needed
(80 char limit), the help text, and updated URL for jacobian

Change-Id: I3b7cce90f72349a986a5b055fa59e53c089ba308