Fix volume type 'is_public' flag update
Update 'is_public' flag only if flag is passed as an argument
while updating volume type. With this change If 'is_public' flag
is not passed by the user, its value will not be updated and
only arguments passed to the update api will be updated.

Added code to check if none of the argument is passed to update
volume-type then it raises the CommandError exception with
appropriate error message.

Change-Id: Ice52e204ebea5d35f04455e74e16745a8bcce3d4
Closes-Bug: #1548708