Remove pypy from tox environment list
Pypy is a supported python interpretor for python-cinderclient,
but I don't think it's necessary to run all unit tests with it
for everyone. We do have the gate-python-cinderclient-pypy test
to validate there are no incompatibilities.

Pypy is also not available by default, so it is an extra step for
someone to know to install and set up the pypy environment before
they can run tox without failures.

Change-Id: I15e315415574dfe8cf8a3d5f7f968a082b81b165