Add Constraints support
Adding constraints support to clients is slightly more complex than
services as the clients themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove python-cinderclient.

Change-Id: Iac0c8a6bcfa9a74a2e836d156642b1e0abf5ec31