Fix adding non-ascii attrs to Resource objects error
Due to these lines of code [0] we don't have an opportunity
to add attributes with non-ascii symbols to Resource objects,
but information about it will be in _info dict.

Example of side effect - quota_show command.
Because we don't have such an attr, here [1] it will be added
None value instead of real value [2].

This patch fixes this problem.

[0] - https://github.com/openstack/python-cinderclient/blob/f8c93ed03b388612ca28b8055debf915ce631cec/cinderclient/apiclient/base.py#L498-L499
[1] - https://github.com/openstack/python-cinderclient/blob/f8c93ed03b388612ca28b8055debf915ce631cec/cinderclient/shell_utils.py#L179
[2] - http://paste.openstack.org/show/593358/

Change-Id: I0493845dafc5dad836e899b9c22d563023c1dab0
Closes-Bug: #1652605