Fix test_auth_with_keystone_v3 test
Test is subject to spurious errors due to an incorrect data check.

The check is assuming that a call to json.dumps with different dicts
will always generate the same string, which is incorrect.

This patch tests the JSON data that is sent in the request on its own
based on converting the passed JSON string to a dict and comparing
expected and actual dicts instead of strings.

TrivialFix

Closes-Bug: #1658704
Change-Id: I386cfee2e2c1dc2971d8a760b485505a90f6f120