Fix attachment_id returned by create and show volume
The attachment_ids in the volume info returned by show volume
were incorrect. It was showing the volume_id, not the actual
attachment_id. This fix changes the attachment_ids returned by show
volume to correctly reflect the attachment_id.

Also, added a unit test for this, and added an attachment_id
to the fake volume so that other tests wouldn't now fail.

Closes-Bug: #1713082
Change-Id: I9ec36af5dd460d03d786aeeb3cc36a869c19ff62