Implement UserID in snapshot list response
We have a gap in the microversions implemented in the cinderclient
and the MAX version being reported by the client.

One of those is the addition of "user_id" in the snapshot list
response.  This patch adds that with the appropriate version
checking.  Note that the Max version has already been updated,
but there are a few implementations missing.

Change-Id: Ia8bdb46b46091a89996e404825466430b8906d18
Partial-Bug: #1715759