Remove setting of version/release from releasenotes
Release notes are version independent, so remove version/release values.
We've found that most projects now require the service package to be
installed in order to build release notes, and this is entirely due to
the current convention of pulling in the version information.

Release notes should not need installation in order to build, so this
unnecessary version setting needs to be removed.

Change-Id: Ia207f4c8d2e6134aed4dbee37ccfb7b7f58bab16
Needed-by: I56909152975f731a9d2c21b2825b972195e48ee8