Removed unnecessary parameters from group and group_snapshots create APIs
As per code, the 'status', 'user_id' and 'project_id' parameter
is not required to be passed in the request body in case of group
and group_snapshot create APIs. Even if you pass these parameter,
it is silently ignored in the code.

This patch removes those parameters passed in the request body.

Change-Id: I29e7d4c8a3eee52f4ea7278d2edf2c7deec40628