Add bindep.txt for system packages
Test jobs will look for a bindep.txt file in the repo to determine
which system packages need to be installed. If it does not find a
local one in the repo it will fall back to an OpenStack-wide bindep
that has many packages not needed for most projects. This adds a
limited list needed for this repo.

For more details, see:
http://lists.openstack.org/pipermail/openstack-dev/2016-August/101590.html

Change-Id: Ic7e00b6a34c0e46c678e5c10ecf0e76faa9913c8