Update python usage docs
The keystoneauth1 get_plugin_loader method requires that a
OS_USER_DOMAIN be specified.  Our docs don't include that
detail.

This change adds the extra needed field.  There's also some
problems currently with loading the client directly as it's
currently documented, which appears to be a bug in the client
code itself.

Change-Id: Idacbb47f9aae1e461379eec4a5e224b0157a8dc4
Partial-Bug: #1758104