Raise API max version for Stein updates
Functionality was added during stein, but the MAX_VERSION of the client
was not updated for these versions. This raises the version to support
the added functionality.

3.57 - Improve volume transfer records - no client changes
3.58 - Add project_id to group - no client changes

3.59 is not included here as that adds sort and pagination that will
need to be slightly reworked before we expose it.

Change-Id: If8a3b88b9a2811ae6681e52dbca62b906d3542eb
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit eae7d55a07c7d1237614e65a6c32a8c6608318ff)