Support backup-restore to a specific volume type or AZ
Enhance the 'backup-restore' shell command to support restoring a
backup to a newly created volume of a specific volume type and/or in a
different AZ. New '--volume-type' and '--availability-zone' arguments
leverage the existing cinder API's ability to create a volume from a
backup, which was added in microversion v3.47.

The shell code is a new v3 implementation, and it drops support for the
v2 command's deprecated '--volume-id' argument.

Change-Id: Ic6645d3b973f8487903c5f57e936ba3b4b3bf005