Add MV 3.63 to the max supported version
During the implementation on [1], it was requested to extend the Cinder
client to support the newly created microversion (MV). Therefore, this
patch is doing exactly that.

We do not need to do any other work forward bcause the list volumes
will display whatever we have in the response. I mean, the default is
``['ID', 'Status', 'Name', 'Size', 'Volume Type', 'Bootable',
'Attached to']``, but one can customize these fields using ``--fields``

[1] https://review.opendev.org/c/openstack/cinder/+/666886

Closes-Bug: https://bugs.launchpad.net/python-cinderclient/+bug/1913474
Depends-on: https://review.opendev.org/c/openstack/cinder/+/666886

Change-Id: I0815570a50e9b38fe18733c727acd52a406bfc1e