[stable-only] Add missing classes to cinderclient.v3
The QuotaSet class was never added to cinderclient.v3.quotas, and
the VolumeTransfer class was never added to cinderclient.v3.
volume_transfers.

Note: this fix is proposed directly to the stable branches, as the
issue was fixed differently in master (Xena development) when the
cinderclient.v2 classes were removed by change I335db5c1799e.

Change-Id: Ie00f068d7818e20683a0c242b51027d657382dd4
Closes-bug: #1938451