Unset tempest.lib timeout in functional tests
The test_cli.CinderBackupTests.test_backup_create_and_delete test
is hitting timeout errors in the python-cinderclient-functional-py36
zuul job.  This is happening because it's inheriting the
OS_TEST_TIMEOUT value of 60 from the base testenv, and that value is
being used by the tempest.lib class we inherit from as a timeout for
each test.  This is a problem for test_backup_create_and_delete
because it creates a volume, waits for available, creates a backup,
waits for available, deletes the volume, waits for deletion, deletes
the backup, waits for deletion.

Our functional tests have their own timeout handling, so turn off
the tempest.lib timeout and use ours.  An alternative to turning it
off is to set it at a value that respects our timeout for our longest
test, which would be:
- time-to-available: 120 sec (x2)
- time-to-deleted: 60 sec (x2)
that is, 360 sec.

Change-Id: I33399b4c094af2cc059da6e332f4c0a91e6ab57e
(cherry picked from commit d04ded6a6ff4cc9ba8377a1b025493e939047b2c)