Correct "Increase default quotas for zuul jobs"
It turns out that you can't use the devstack post_config phase
for this because it happens too late.  The quota settings must
be made to cinder.conf before cinder-manage db_sync is called
where they're used to set the max limit for the tables used by
the db quotas driver.

Depends-on: https://review.opendev.org/c/openstack/devstack/+/803521
Change-Id: I3bb7d9f9bede063573b24fb5a33a364c53d52434