[stable-only] Skip backup test in functional py39 job
python-cinderclient-functional-py39 job has been failing since
23 November, 2022[1]. The failure has been described in detail
in the mailing thread replies[2][3].
In summary, a new iSCSI target path was added in the kernel,
/sys/kernel/config/target/iscsi/*cpus_allowed_list*, which
is causing rtslib-fb 2.7.4 to fail. This issue was fixed in
rtslib-fb 2.7.5 but the upper-constraints in stable/yoga branch
is 2.7.4[4]. This isn't a problem in stable/zed since the version
used for rtslib-fb is 2.7.5[5].

I have proposed a requirements patch to bump the rtslib-fb version
to 2.7.5[6] and tested it against cinderclient[7] where the
python-cinderclient-functional-py39 job is passing.

I've tried reaching out to the requirements team on IRC and on ML
to make this change but there hasn't been any response.
The best way to move forward and unblock our gate is to skip the
failing test test_backup_create_and_delete and unblock changes
in the stable/yoga branch.

[1] https://zuul.openstack.org/builds?job_name=python-cinderclient-functional-py39&branch=stable%2Fyoga&skip=0
[2] https://lists.openstack.org/pipermail/openstack-discuss/2023-January/031947.html
[3] https://lists.openstack.org/pipermail/openstack-discuss/2023-January/031958.html
[4] https://opendev.org/openstack/requirements/src/branch/stable/yoga/upper-constraints.txt#L13
[5] https://opendev.org/openstack/requirements/src/branch/stable/zed/upper-constraints.txt#L13
[6] https://review.opendev.org/c/openstack/requirements/+/870714
[7] https://review.opendev.org/c/openstack/python-cinderclient/+/870513

Closes-Bug: #2008010

Change-Id: I54b1d797d9662ba180e146ee91ea7b5ed6d0a0a1