Use generic testing template
Change If402f9ae0ca06fec0 replaced cycle-specific testing templates
that had to be changed in each project's zuul config file with a
generic template that only needs to be updated in one place, namely,
in the openstack-zuul-jobs repo.

Apparently cinderclient didn't get the memo, so we fix that now.

Change-Id: I1144ed99b98e91035c88bc4cc8a065ab0249a012