Honour all_tenants in consistency group list
Although consistency groups are deprecated, the CLI command and
API support still exists. Currently it accepts the --all-tenant
parameter but doesn't pass it to the API request.
This patch corrects that behavior.

The test for consistency group were removed with the removal of
v2 API and also no tests existed for consistency group list so
no point in adding them at this point.

Change-Id: Ib3557efa50941d75d7f7f0cac01404f5c2db4526