[CI][stable-only] Replace centos-8 nodeset and fix gate
This patch fixes 3 problems to unblock the gate:

1. Replace centos-8 nodeset with ubuntu-bionic

devstack-single-node-centos-8-stream nodeset was dropper recently from
zuul config, so now Zuul shows config error for jobs that were using
this nodeset. centos 8 was used to be able to install py36 for testing
the minimum python version that can be used in Yoga and older branches.

This patch replaces to ubuntu-bionic as that is still available and
that has py36 by default.
Note, that this unfortunately failing because bionic is not supported
in this version in devstack and needs a FORCE=yes entry in local.conf.
So this job needs to be set as non-voting temporarily and set back to
voting as soon as the job is fixed.

2. Cap setuptools <71.0.0

Furthermore, this change caps setuptools to be <71.0.0 as py39 docs
job (on ubuntu-focal) started to fail due to recent virtualenv
release (20.26.4) on Yoga (which bundles setuptools), because we have
'packaging==21.3' in this branch that is not compatible with newer
setuptools [1].

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

3. Set py39 based functional job as non-voting

This is needed as py39 job on CentOS Stream 9 fails as there is no
pip installed. Similarly like with the py36 functional job, this
needs to be set back to voting as soon as the job is fixed.

[1] https://github.com/pypa/setuptools/issues/4483

Change-Id: I007be9e8b99060abe1756396603871fdc50a50fe
Signed-off-by: Elod Illes <elod.illes@est.tech>